pyproject.toml
[tool.poetry]
name = "youtudownload-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyrogram = "^2.0.106"
python-dotenv = "^1.0.1"
tgcrypto = "^1.2.5"
requests = "^2.31.0"
pillow = "^10.2.0"
yt-dlp = "^2024.3.10"
numpy = "^1.26.4"
moviepy = "^1.0.3"
aiohttp = "^3.9.3"
jinja2 = "^3.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
